+++ front/lib/Screens/home/basket/spare_part_details_page.dart
