+++ front/lib/Screens/home/cart/CartPage.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'CheckoutPage.dart';
 
 class CartPage extends StatefulWidget {
