+++ front/lib/Screens/home/product/productpage.dart
@@ -1,8 +1,6 @@
-// ignore_for_file: library_private_types_in_public_api
-
 import 'package:flutter/material.dart';
 import 'productdetails.dart';
-import 'cartpage.dart';
+import '../cart/CartPage.dart';
 
 class ProductPage extends StatefulWidget {
   final String categoryName;
