+++ front/lib/Screens/owner/owner.dart
@@ -1,3 +1,4 @@
+// this page for compony owner (grage owner)
 import 'package:flutter/material.dart';
 import 'porfile.dart';
 import 'employee.dart';
